# FCN5 update for single and multi-GPU performance