--- a/tools/tensorflow/fc/fcn5_mnist.py
+++ b/tools/tensorflow/fc/fcn5_mnist.py
@@ -22,11 +22,15 @@ tf.app.flags.DEFINE_boolean('use_fp16', False,
                             """Train the model using fp16.""")
 tf.app.flags.DEFINE_boolean('log_device_placement', False,
                             """Whether to log device placement.""")
+# DataSets is included with TensorFlow 1.2 and is the easiest way to create input pipelines.
+# It is faster than feed_dict 1 GPU and is much faster at 2,4 and 8 GPUs.
+tf.app.flags.DEFINE_boolean('use_dataset', False,
+                            """Whether to use datasets vs. feed_dict.""")
 tf.app.flags.DEFINE_integer('num_gpus', 1, """How many GPUs to use.""")
+tf.app.flags.DEFINE_boolean('xla', False,
+                            """True to use XLA, which has to be compiled in.""")
 
 EPOCH_SIZE = 60000
-TEST_SIZE = 10000
-
 
 def createFakeData(count, featureDim, labelDim):
     features = np.random.randn(count, featureDim)
@@ -65,15 +69,38 @@ def train(model='fcn5'):
         device_str = '/cpu:0'
         num_threads = os.getenv('OMP_NUM_THREADS', 1)
         config = tf.ConfigProto(allow_soft_placement=True, intra_op_parallelism_threads=int(num_threads))
-
-
+    
+    if FLAGS.xla:
+        # Turns on XLA.  XLA is not included in the standard build.  For single GPU this shows ~5% improvement
+        config.graph_options.optimizer_options.global_jit_level = tf.OptimizerOptions.ON_1
+    
     with tf.Graph().as_default(), tf.device(device_str), tf.Session(config=config) as sess:
         feature_dim = models.feature_dim
         label_dim = models.label_dim
-        images = tf.placeholder(tf.float32, [None, feature_dim])
-        labels = tf.placeholder(tf.float32, [None, label_dim])
+        images = None
+        labels = None
+        iterator = None
+        if FLAGS.use_dataset:
+            with tf.device('/CPU:0'):
+                d_features = mnist.train.images
+                d_labels = mnist.train.labels
+                dataset = tf.contrib.data.Dataset.from_tensor_slices((d_features, d_labels))
+                dataset = dataset.repeat()
+                dataset = dataset.shuffle(buffer_size=60000)
+                dataset = dataset.batch(FLAGS.batch_size)
+                # Trick to get datasets to buffer the next epoch.  This is needed because
+                # the data loading is occuring outside DataSets in python.  Normally preprocessing
+                # would occur in DataSets and this odd looking line is not needed.  
+                dataset = dataset.map(lambda x,y:(x,y),num_threads=1,output_buffer_size=1)
+                iterator = dataset.make_initializable_iterator()
+                images,labels = iterator.get_next()
+                
+        else:
+            images = tf.placeholder(tf.float32, [None, feature_dim], name="images_placeholder")
+            labels = tf.placeholder(tf.int64, [None, label_dim], name="labels_placeholder")
 
         logits = None
+        loss = None
         if model == 'fcn5':
             logits = models.model_fcn5(images)
         else:
@@ -84,23 +111,29 @@ def train(model='fcn5'):
         accuracy = tf.reduce_mean(tf.cast(predictionCorrectness, "float"))
 
         lr = 0.05
-        #optimizer = tf.train.GradientDescentOptimizer(lr).minimize(loss)
         optimizer = tf.train.MomentumOptimizer(lr, 0.9).minimize(loss)
 
         init = tf.global_variables_initializer()
+
         sess.run(init)
-        tf.train.start_queue_runners(sess=sess)
+        if FLAGS.use_dataset:
+            sess.run(iterator.initializer)
         batch_size_per_epoch = int((EPOCH_SIZE + FLAGS.batch_size - 1)/ FLAGS.batch_size)
-        iterations = FLAGS.epochs * batch_size_per_epoch 
+        iterations = FLAGS.epochs * batch_size_per_epoch
         average_batch_time = 0.0
         epochs_info = []
         average_loss = 0.0
         for step in range(iterations):
             start_time = time.time()
-            imgs, labs = get_real_batch_data(FLAGS.batch_size, 10)
-            _, loss_value = sess.run([optimizer, loss], feed_dict={images:imgs,labels:labs})
-            average_loss += loss_value
+            imgs = None
+            labs = None
+            if FLAGS.use_dataset:
+                _, loss_value = sess.run([optimizer, loss])
+            else:
+                imgs, labs = get_real_batch_data(FLAGS.batch_size, 10)
+                _, loss_value = sess.run([optimizer, loss], feed_dict={images:imgs,labels:labs})
             duration = time.time() - start_time
+            average_loss += loss_value
             average_batch_time += float(duration)
             assert not np.isnan(loss_value), 'Model diverged with loss = NaN'
             if step % FLAGS.log_step == 0:
@@ -121,6 +154,7 @@ def train(model='fcn5'):
 
 def main(argv=None):
     os.environ['TF_SYNC_ON_FINISH'] = '0'
+    os.environ['TF_ENABLE_WINOGRAD_NONFUSED'] = '1'
     train(model='fcn5')