@shyhuai 2017-05-29
Hi, @tfboyd , many thanks for your contribution. The core frequency of the K80 GPU is kept at the default core clock with 562MHz, and the auto boost is off. The two K80s on a tested machine are not peered, which may result in the performance decrease if one doesn't perform well at data synchronization on multi-gpu environments. We will first release the results of your codes that work properly on tf1.0(4ac9c09) since our v8 version will not update to tf1.2. Thank you!