@FreemanX 2017-05-29
Thank you! We tested on our K80 with frequency settings being 562MHz for sm and 2505MHz for memory. 
With use_datasets=False, xla=False, batch_size=1024, we get around 60ms/batch.