@tfboyd 2017-05-29
60ms is not far from my test so that is good and confirms that my setup is not far from yours.  I have alexnet changes where I move to NCHW I will submit tomorrow.