@tfboyd 2017-05-29
@FreemanX   

I set the defaults to work with TF 1.0 or 1.1.  So you do not need to change any of your scripts.  I was not remotely asking you to run on 1.2.  I am doing the same with my other Pull Requests.